requirements.txt
py==1.10.0
pytest==6.2.5
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-xdist==2.4.0
requests==2.26.0
python-dotenv==0.19.1